refactor(mediawiki): split write.go into files/move/categories siblings (deslop #59)
