+++ wiki/write.go
@@ -3,12 +3,9 @@ package wiki
 import (
 	"context"
 	"fmt"
-	"io"
-	"net/http"
 	"net/url"
 	"regexp"
 	"strings"
-	"time"
 )
 
 // EditPage creates or edits a page
@@ -486,499 +483,4 @@ func (c *Client) buildEditRevisionInfo(title string, oldRevision, newRevision in
 // response and returns a map of normalized→original title. MediaWiki normalizes
 // titles before lookup, so the original input title may not appear in the
 // "pages" map.
-func extractNormalizedTitleMap(query map[string]interface{}) map[string]string {
-	normalized := make(map[string]string)
-	normList, ok := query["normalized"].([]interface{})
-	if !ok {
-		return normalized
-	}
-	for _, n := range normList {
-		norm, ok := n.(map[string]interface{})
-		if !ok {
-			continue
-		}
-		normalized[getString(norm["to"])] = getString(norm["from"])
-	}
-	return normalized
-}
-
-// recordExistenceFromBatchPages walks the "pages" map from a query response
-// and writes existence flags into result, including the un-normalized form
-// if applicable.
-func recordExistenceFromBatchPages(pages map[string]interface{}, normalized map[string]string, result map[string]bool) {
-	for _, pageData := range pages {
-		page, ok := pageData.(map[string]interface{})
-		if !ok {
-			continue
-		}
-		title := getString(page["title"])
-		_, missing := page["missing"]
-		exists := !missing
-		result[title] = exists
-		if originalTitle, ok := normalized[title]; ok {
-			result[originalTitle] = exists
-		}
-	}
-}
-
-// queryPageExistence runs one batch existence query and merges the outcome
-// into result. Returns the API error (network/auth) without aborting on
-// missing-from-response cases — those are caught by the caller's fallback loop.
-func (c *Client) queryPageExistence(ctx context.Context, batch []string, result map[string]bool) error {
-	params := url.Values{}
-	params.Set("action", "query")
-	params.Set("titles", strings.Join(batch, "|"))
-	resp, err := c.apiRequest(ctx, params)
-	if err != nil {
-		return err
-	}
-	query, ok := resp["query"].(map[string]interface{})
-	if !ok {
-		return nil
-	}
-	pages, ok := query["pages"].(map[string]interface{})
-	if !ok {
-		return nil
-	}
-	recordExistenceFromBatchPages(pages, extractNormalizedTitleMap(query), result)
-	return nil
-}
-
-// checkPagesExist checks if multiple pages exist using MediaWiki's multi-value API
-// Returns a map of page title -> exists (bool)
-// This is much more efficient than calling GetPageInfo for each page individually
-func (c *Client) checkPagesExist(ctx context.Context, titles []string) (map[string]bool, error) {
-	if len(titles) == 0 {
-		return make(map[string]bool), nil
-	}
-
-	const maxTitlesPerRequest = 50
-	result := make(map[string]bool, len(titles))
-
-	for i := 0; i < len(titles); i += maxTitlesPerRequest {
-		end := i + maxTitlesPerRequest
-		if end > len(titles) {
-			end = len(titles)
-		}
-		if err := c.queryPageExistence(ctx, titles[i:end], result); err != nil {
-			return nil, err
-		}
-	}
-
-	// Default any titles that didn't appear in any response to non-existent.
-	for _, title := range titles {
-		if _, ok := result[title]; !ok {
-			result[title] = false
-		}
-	}
-	return result, nil
-}
-
-// fileTypeFromMIME returns a friendly file type label from a MIME string.
-func fileTypeFromMIME(mimeType string) string {
-	if mimeType == "application/pdf" {
-		return "pdf"
-	}
-	if strings.HasPrefix(mimeType, "text/") {
-		return strings.TrimPrefix(mimeType, "text/")
-	}
-	return mimeType
-}
-
-// extractFileURLFromPageEntry pulls the download URL and file type from one
-// entry in the imageinfo response. Returns an error if the entry is malformed,
-// represents a missing file, or has no URL.
-func extractFileURLFromPageEntry(page map[string]interface{}, filename string) (fileURL, fileType string, err error) {
-	if _, missing := page["missing"]; missing {
-		return "", "", fmt.Errorf("file '%s' does not exist", filename)
-	}
-	imageinfo, ok := page["imageinfo"].([]interface{})
-	if !ok || len(imageinfo) == 0 {
-		return "", "", fmt.Errorf("no file info available for '%s'", filename)
-	}
-	info, ok := imageinfo[0].(map[string]interface{})
-	if !ok {
-		return "", "", fmt.Errorf("unexpected API response: invalid imageinfo format for '%s'", filename)
-	}
-	fileURL = getString(info["url"])
-	if fileURL == "" {
-		return "", "", fmt.Errorf("no download URL for '%s'", filename)
-	}
-	return fileURL, fileTypeFromMIME(getString(info["mime"])), nil
-}
-
-func (c *Client) getFileURL(ctx context.Context, filename string) (string, string, error) {
-	params := url.Values{}
-	params.Set("action", "query")
-	params.Set("titles", filename)
-	params.Set("prop", "imageinfo")
-	params.Set("iiprop", "url|mime|size")
-
-	resp, err := c.apiRequest(ctx, params)
-	if err != nil {
-		return "", "", err
-	}
-	query, ok := resp["query"].(map[string]interface{})
-	if !ok {
-		return "", "", fmt.Errorf("unexpected response format")
-	}
-	pages, ok := query["pages"].(map[string]interface{})
-	if !ok {
-		return "", "", fmt.Errorf("no pages in response")
-	}
-
-	for _, pageData := range pages {
-		page, ok := pageData.(map[string]interface{})
-		if !ok {
-			continue
-		}
-		return extractFileURLFromPageEntry(page, filename)
-	}
-	return "", "", fmt.Errorf("file '%s' not found", filename)
-}
-
-// downloadFile downloads a file from the given URL.
-//
-// SECURITY: The fileURL is validated against SSRF (private IPs, DNS rebinding,
-// redirect bypass) via validateFileURL before any network I/O. The actual
-// HTTP request goes through downloadClient, which uses safeDialer +
-// CheckRedirect for defense in depth. Callers in production today (only
-// SearchInFile) get fileURL from the wiki's own imageinfo API, so the URL is
-// already trusted. The validation closes the gap for any future caller that
-// might accept user-supplied URLs.
-//
-// Tests can set c.allowPrivateDownloadForTest to skip validation against
-// httptest servers (which bind to 127.0.0.1).
-func (c *Client) downloadFile(ctx context.Context, fileURL string) ([]byte, error) {
-	httpClient := downloadClient
-	if !c.allowPrivateDownloadForTest {
-		if err := validateFileURL(fileURL); err != nil {
-			return nil, fmt.Errorf("download URL rejected: %w", err)
-		}
-	} else {
-		// Test path: bypass SSRF guards so httptest (loopback) servers work.
-		httpClient = c.httpClient
-	}
-
-	req, err := http.NewRequestWithContext(ctx, "GET", fileURL, nil)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create request: %w", err)
-	}
-
-	req.Header.Set("User-Agent", c.config.UserAgent)
-
-	// Forward session cookies from the API client's jar onto the download
-	// request. Required for MediaWiki instances that gate file access behind
-	// a login (the $wgUploadDirectory + img_auth.php pattern common on
-	// corporate wikis): the file-serving endpoint refuses anonymous reads
-	// even when the bot is authenticated for api.php. The standard cookie jar
-	// is domain-scoped (RFC 6265), so cookies set for the wiki host are only
-	// forwarded when fileURL is on the same host — no credential leakage to
-	// arbitrary external hosts.
-	if c.httpClient != nil && c.httpClient.Jar != nil {
-		for _, cookie := range c.httpClient.Jar.Cookies(req.URL) {
-			req.AddCookie(cookie)
-		}
-	}
-
-	resp, err := httpClient.Do(req) // #nosec G107 G704 -- URL validated by validateFileURL; request goes through downloadClient (safeDialer + CheckRedirect)
-	if err != nil {
-		return nil, fmt.Errorf("failed to download: %w", err)
-	}
-	defer func() { _ = resp.Body.Close() }()
-
-	if resp.StatusCode != http.StatusOK {
-		// Surface the known bot-password + img_auth.php gotcha with an
-		// actionable message rather than a bare status code. Wikis that gate
-		// file access via img_auth.php check user-level auth (UserID/UserName/
-		// Token cookies), which the legacy `action=login` bot-password flow
-		// doesn't establish — only a BPsession cookie. The session is fine
-		// for api.php but the file-serving endpoint rejects it. Tracked as
-		// mediawiki-mcp-server-xum.
-		if resp.StatusCode == http.StatusForbidden &&
-			c.config.HasCredentials() &&
-			strings.Contains(req.URL.Path, "/img_auth.php") {
-			return nil, fmt.Errorf(
-				"download failed with status 403: %s is gated behind img_auth.php and refused the bot-password session. "+
-					"Bot-password logins authenticate against api.php but img_auth.php enforces user-level auth this flow does not establish. "+
-					"This is a wiki-side limitation, not a CLI bug. Workarounds: download the file manually from the wiki UI, "+
-					"or ask a wiki admin to grant additional rights to your bot password at Special:BotPasswords",
-				req.URL.Host)
-		}
-		return nil, fmt.Errorf("download failed with status %d", resp.StatusCode)
-	}
-
-	// Limit download size to 50MB
-	const maxSize = 50 * 1024 * 1024
-	limitedReader := &io.LimitedReader{R: resp.Body, N: maxSize}
-
-	data, err := io.ReadAll(limitedReader)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read file data: %w", err)
-	}
-
-	if limitedReader.N <= 0 {
-		return nil, fmt.Errorf("file exceeds maximum size of 50MB")
-	}
-
-	return data, nil
-}
-
 // performMove executes a single move attempt with a fresh CSRF token.
-func (c *Client) performMove(ctx context.Context, args MovePageArgs) (map[string]interface{}, error) {
-	token, err := c.getCSRFToken(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("authentication failed: %w", err)
-	}
-
-	params := url.Values{}
-	params.Set("action", "move")
-	params.Set("from", args.From)
-	params.Set("to", args.To)
-	params.Set("token", token)
-
-	if args.Reason != "" {
-		params.Set("reason", args.Reason)
-	}
-
-	if args.NoRedirect {
-		params.Set("noredirect", "1")
-	}
-
-	// Default: move talk page
-	if !args.MoveTalk {
-		// MediaWiki moves talk by default, so we only set movetalk=0 if explicitly disabled
-	} else {
-		params.Set("movetalk", "1")
-	}
-
-	if args.MoveSubpages {
-		params.Set("movesubpages", "1")
-	}
-
-	resp, err := c.apiRequest(ctx, params)
-	if err != nil {
-		return nil, err
-	}
-
-	// Check for badtoken error so caller can retry
-	if errInfo, ok := resp["error"].(map[string]interface{}); ok {
-		code := getString(errInfo["code"])
-		if code == "badtoken" {
-			return nil, fmt.Errorf("%s: %s", code, getString(errInfo["info"]))
-		}
-	}
-
-	return resp, nil
-}
-
-// MovePage moves (renames) a wiki page
-func (c *Client) MovePage(ctx context.Context, args MovePageArgs) (MovePageResult, error) {
-	if args.From == "" {
-		return MovePageResult{}, &ValidationError{
-			Field:   "from",
-			Message: "source page title is required",
-		}
-	}
-	if args.To == "" {
-		return MovePageResult{}, &ValidationError{
-			Field:   "to",
-			Message: "target page title is required",
-		}
-	}
-
-	if err := c.EnsureLoggedIn(ctx); err != nil {
-		return MovePageResult{}, fmt.Errorf("authentication required for page moves: %w", err)
-	}
-
-	resp, err := c.performMove(ctx, args)
-	if err != nil && strings.Contains(err.Error(), "badtoken") {
-		c.invalidateCSRFToken()
-		resp, err = c.performMove(ctx, args)
-	}
-	if err != nil {
-		return MovePageResult{}, err
-	}
-
-	// Check for errors
-	if errInfo, ok := resp["error"].(map[string]interface{}); ok {
-		return MovePageResult{
-			Success: false,
-			From:    args.From,
-			To:      args.To,
-			Message: fmt.Sprintf("Move failed: %s", getString(errInfo["info"])),
-		}, nil
-	}
-
-	moveData, ok := resp["move"].(map[string]interface{})
-	if !ok {
-		return MovePageResult{
-			Success: false,
-			From:    args.From,
-			To:      args.To,
-			Message: "Unexpected response format",
-		}, nil
-	}
-
-	result := MovePageResult{
-		Success: true,
-		From:    getString(moveData["from"]),
-		To:      getString(moveData["to"]),
-		Reason:  args.Reason,
-		Message: fmt.Sprintf("Page moved from '%s' to '%s'", getString(moveData["from"]), getString(moveData["to"])),
-	}
-
-	// Check if talk page was moved
-	if _, hasTalkFrom := moveData["talkfrom"]; hasTalkFrom {
-		result.TalkMoved = true
-	}
-
-	// Build redirect URL
-	wikiBaseURL := strings.TrimSuffix(c.config.BaseURL, "api.php") + "index.php"
-	encodedFrom := url.QueryEscape(strings.ReplaceAll(result.From, " ", "_"))
-	result.RedirectURL = fmt.Sprintf("%s?title=%s&redirect=no", wikiBaseURL, encodedFrom)
-
-	// Log the move
-	c.logAudit(AuditEntry{
-		Timestamp: time.Now().UTC().Format(time.RFC3339),
-		Operation: "move",
-		Title:     result.From + " → " + result.To,
-		Summary:   args.Reason,
-		WikiURL:   c.config.BaseURL,
-		Success:   true,
-	})
-
-	return result, nil
-}
-
-// ManageCategories adds or removes categories from a page
-// categoryTagRegex matches "[[Category:Name|sortkey]]" anywhere in wikitext.
-var categoryTagRegex = regexp.MustCompile(`\[\[Category:([^\]|]+)(?:\|[^\]]*)?\]\]`)
-
-// parseExistingCategories returns the set of category names already present in
-// the wikitext content.
-func parseExistingCategories(content string) map[string]bool {
-	existing := make(map[string]bool)
-	for _, m := range categoryTagRegex.FindAllStringSubmatch(content, -1) {
-		existing[strings.TrimSpace(m[1])] = true
-	}
-	return existing
-}
-
-// removeCategoriesFromContent removes the listed categories from the content
-// and returns the rewritten content plus the per-category outcome (removed,
-// not-found). The existing-set is updated in place.
-func removeCategoriesFromContent(content string, toRemove []string, existing map[string]bool) (newContent string, removed, notFound []string) {
-	newContent = content
-	for _, cat := range toRemove {
-		cat = strings.TrimSpace(cat)
-		if !existing[cat] {
-			notFound = append(notFound, cat)
-			continue
-		}
-		removeRegex := regexp.MustCompile(`\n?\[\[Category:` + regexp.QuoteMeta(cat) + `(?:\|[^\]]*)?\]\]\n?`)
-		newContent = removeRegex.ReplaceAllString(newContent, "\n")
-		removed = append(removed, cat)
-		delete(existing, cat)
-	}
-	return newContent, removed, notFound
-}
-
-// addCategoriesToContent appends category tags missing from the existing-set.
-// Categories already present are reported via alreadyPresent.
-func addCategoriesToContent(content string, toAdd []string, existing map[string]bool) (newContent string, added, alreadyPresent []string) {
-	newContent = content
-	for _, cat := range toAdd {
-		cat = strings.TrimSpace(cat)
-		if existing[cat] {
-			alreadyPresent = append(alreadyPresent, cat)
-			continue
-		}
-		newContent = strings.TrimRight(newContent, "\n") + "\n[[Category:" + cat + "]]\n"
-		added = append(added, cat)
-		existing[cat] = true
-	}
-	return newContent, added, alreadyPresent
-}
-
-// buildCategoryEditSummary composes the default edit summary for category changes.
-func buildCategoryEditSummary(added, removed []string) string {
-	parts := make([]string, 0, 2)
-	if len(added) > 0 {
-		parts = append(parts, fmt.Sprintf("Added categories: %s", strings.Join(added, ", ")))
-	}
-	if len(removed) > 0 {
-		parts = append(parts, fmt.Sprintf("Removed categories: %s", strings.Join(removed, ", ")))
-	}
-	return strings.Join(parts, ". ")
-}
-
-func (c *Client) ManageCategories(ctx context.Context, args ManageCategoriesArgs) (ManageCategoriesResult, error) {
-	if args.Title == "" {
-		return ManageCategoriesResult{}, fmt.Errorf("title is required")
-	}
-	if len(args.Add) == 0 && len(args.Remove) == 0 {
-		return ManageCategoriesResult{}, fmt.Errorf("at least one category to add or remove is required")
-	}
-
-	page, err := c.GetPage(ctx, GetPageArgs{Title: args.Title, Format: "wikitext"})
-	if err != nil {
-		return ManageCategoriesResult{}, fmt.Errorf("failed to get page: %w", err)
-	}
-
-	existing := parseExistingCategories(page.Content)
-	result := ManageCategoriesResult{
-		Title:             page.Title,
-		Preview:           args.Preview,
-		CurrentCategories: keysOf(existing),
-	}
-
-	newContent, removed, notFound := removeCategoriesFromContent(page.Content, args.Remove, existing)
-	result.Removed = removed
-	result.NotFound = notFound
-
-	newContent, added, alreadyPresent := addCategoriesToContent(newContent, args.Add, existing)
-	result.Added = added
-	result.AlreadyPresent = alreadyPresent
-	result.CurrentCategories = keysOf(existing)
-
-	if len(result.Added) == 0 && len(result.Removed) == 0 {
-		result.Success = true
-		result.Message = "No changes needed"
-		return result, nil
-	}
-	if args.Preview {
-		result.Success = true
-		result.Message = fmt.Sprintf("Preview: would add %d and remove %d categories", len(result.Added), len(result.Removed))
-		return result, nil
-	}
-
-	summary := args.Summary
-	if summary == "" {
-		summary = buildCategoryEditSummary(result.Added, result.Removed)
-	}
-	oldRevision := page.Revision
-	editResult, err := c.EditPage(ctx, EditPageArgs{
-		Title:   page.Title,
-		Content: newContent,
-		Summary: summary,
-		Minor:   true,
-	})
-	if err != nil {
-		return ManageCategoriesResult{}, fmt.Errorf("failed to save changes: %w", err)
-	}
-	result.Success = editResult.Success
-	result.RevisionID = editResult.RevisionID
-	result.Message = fmt.Sprintf("Added %d, removed %d categories", len(result.Added), len(result.Removed))
-	result.Revision, result.Undo = c.buildEditRevisionInfo(page.Title, oldRevision, editResult.RevisionID)
-	return result, nil
-}
-
-// keysOf returns the keys of a string-bool map as a slice.
-func keysOf(m map[string]bool) []string {
-	out := make([]string, 0, len(m))
-	for k := range m {
-		out = append(out, k)
-	}
-	return out
-}
